[BE] fix : 상담 예약 정보 조회 분리

See merge request s11-webmobile1-sub2/S11P12B301!121

+++ BE/src/main/java/com/ssafy/inmind/reservation/entity/Reservation.java
@@ -25,14 +25,14 @@ public class Reservation extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idx")
-    private long id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_idx", nullable = false)
     private User user;
 
     @Column(name = "child_idx", nullable = false)
-    private int child;
+    private Long child;
 
     @OneToOne(mappedBy = "reservation", cascade = CascadeType.ALL)
     private Review review;
